+++ ext-src/lib/worker.ts
@@ -1,91 +1,8 @@
-// @ts-ignore
-import * as path from 'path';
 import * as fs from 'fs';
 import * as grpc from '@grpc/grpc-js';
-import * as protoLoader from '@grpc/proto-loader';
-import { sha3 } from '../hash/sha3';
 import { ABIParameter } from '../types';
-
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const EthereumTx = require('ethereumjs-tx').Transaction;
-// eslint-disable-next-line @typescript-eslint/no-var-requires
-const { formatters } = require('web3-core-helpers');
-
-const PROTO_PATH = [path.join(__dirname, '../../services/client-call.proto')];
-const packageDefinition = protoLoader.loadSync(PROTO_PATH, {
-  keepCase: true,
-  longs: String,
-  enums: String,
-  defaults: true,
-  oneofs: true,
-});
-const protoDescriptor = grpc.loadPackageDefinition(packageDefinition) as any;
-
-// client-call grpc
-const clientCallPB = protoDescriptor.eth_client_call;
-let clientCallClient: any;
-try {
-  clientCallClient = new clientCallPB.ClientCallService('cc.ethcode.dev:50053', grpc.credentials.createInsecure());
-  // clientCallClient = new clientCallPB.ClientCallService('192.168.1.116:50053', grpc.credentials.createInsecure());
-} catch (e) {
-  // @ts-ignore
-  process.send({ error: e });
-}
-
-// sign an unsigned raw transaction and deploy
-function deployUnsignedTx(meta: any, tx: any, privateKey: any, testnetId?: any) {
-  try {
-    // eslint-disable-next-line no-param-reassign
-    tx = JSON.parse(tx);
-    const txData = formatters.inputTransactionFormatter(tx);
-    const chainId = Number(testnetId);
-    const unsignedTransaction = new EthereumTx(
-      {
-        from: txData.from || '0x',
-        nonce: txData.nonce || '0x',
-        gasPrice: txData.gasPrice,
-        gas: txData.gas || '0x',
-        to: txData.to || '0x',
-        value: txData.value || '0x',
-        data: txData.data || '0x',
-      },
-      { chain: chainId }
-    );
-    const pvtk = Buffer.from(privateKey, 'hex');
-    unsignedTransaction.sign(pvtk);
-    const rlpEncoded = unsignedTransaction.serialize().toString('hex');
-    const rawTransaction = `0x${rlpEncoded}`;
-    const transactionHash = sha3(rawTransaction);
-    // @ts-ignore
-    process.send({ responses: transactionHash });
-    const c = {
-      callInterface: {
-        command: 'deploy-signed-tx',
-        payload: rawTransaction,
-        testnetId,
-      },
-    };
-
-    const call = clientCallClient.RunDeploy(c, meta, (err: any) => {
-      if (err) {
-        console.error(err);
-      }
-    });
-
-    call.on('data', (data: any) => {
-      // @ts-ignore
-      process.send({ transactionResult: data.result });
-    });
-    call.on('error', (err: Error) => {
-      // @ts-ignore
-      process.send({ error: err });
-    });
-  } catch (error) {
-    console.log(error);
-    // @ts-ignore
-    process.send({ error: error.message });
-  }
-}
+import { clientCallClient } from './proto';
+import { deployUnsignedTx, deployGanacheTx } from './deployUnsignedTransaction';
 
 // create constructor input file
 function writeConstrucor(path: string, inputs: Array<ABIParameter>) {
@@ -206,45 +123,8 @@ process.on('message', async (m) => {
   }
   // Deploy
   if (m.command === 'deploy-contract') {
-    if (m.authToken) {
-      // @ts-ignore
-      process.send({ authToken: m.authToken });
-    }
-    const { from, abi, bytecode, params, gasSupply } = m.payload;
-    const inp = {
-      from,
-      abi,
-      bytecode,
-      params,
-      gasSupply: typeof gasSupply === 'string' ? parseInt(gasSupply, 10) : gasSupply,
-    };
-    const c = {
-      callInterface: {
-        command: 'deploy-contract',
-        payload: JSON.stringify(inp),
-        testnetId: m.testnetId,
-      },
-    };
-    const call = clientCallClient.RunDeploy(c, meta, (err: any, response: any) => {
-      if (err) {
-        // @ts-ignore
-        process.send({ error: err });
-      } else {
-        // @ts-ignore
-        process.send({ response });
-      }
-    });
-    call.on('data', (data: any) => {
-      // @ts-ignore
-      process.send({ deployedResult: data.result });
-    });
-    call.on('end', () => {
-      process.exit(0);
-    });
-    call.on('error', (err: Error) => {
-      // @ts-ignore
-      process.send({ error: err });
-    });
+    const { unsignedTx } = m.payload;
+    deployGanacheTx(meta, unsignedTx, m.testnetId);
   }
   // Method call
   if (m.command === 'ganache-contract-method-call') {
@@ -330,30 +210,23 @@ process.on('message', async (m) => {
   // Gas Estimate
   if (m.command === 'get-gas-estimate') {
     const { abi, bytecode, params, from } = m.payload;
-    const inp = { abi, bytecode, params, from };
     const c = {
-      callInterface: {
-        command: 'get-gas-estimate',
-        payload: JSON.stringify(inp),
-        testnetId: m.testnetId,
-      },
+      networkid: m.testnetId,
+      abi: JSON.stringify(abi),
+      bytecode,
+      params: JSON.stringify(params),
+      address: from,
+      fromAddress: from,
+      value: 0,
     };
-    const call = clientCallClient.RunDeploy(c, meta, (err: any, response: any) => {
+    clientCallClient.EstimateGas(c, meta, (err: any, response: any) => {
       if (err) {
         // @ts-ignore
         process.send({ error: err });
       } else {
         // @ts-ignore
-        process.send({ response });
+        process.send({ gasEstimate: response.result });
       }
-    });
-    call.on('data', (data: any) => {
-      // @ts-ignore
-      process.send({ gasEstimate: data.result });
-    });
-    call.on('error', (err: Error) => {
-      // @ts-ignore
-      process.send({ error: err });
     });
   }
   // Build raw transaction for contract creation
