--- services/remix-debug.proto
@@ -1,22 +1,0 @@
-syntax = "proto3";
-
-package remix_debug;
-option go_package = "remix_debug";
-
-message DebugInterface {
-    string command = 1;
-    string payload = 2;
-    string testnetId = 3;
-}
-
-message DebugRequest {
-    DebugInterface debugInterface = 1;
-}
-
-message DebugResponse {
-    string result = 1;
-}
-
-service RemixDebugService {
-    rpc RunDebug(DebugRequest) returns (stream DebugResponse) {};
-}
