--- services/remix-tests.proto
@@ -1,21 +1,0 @@
-syntax = "proto3";
-
-package remix_tests;
-option go_package = "remix_tests";
-
-message TestInterface {
-    string command = 1;
-    string payload = 2;
-}
-
-message TestRequest {
-    TestInterface testInterface = 1;
-}
-
-message TestResponse {
-    string result = 1;
-}
-
-service RemixTestsService {
-    rpc RunTests(TestRequest) returns (stream TestResponse) {};
-}
